[REFACTOR] move modal to the component
