+++ src/components/game-over-modal/game-over-modal.component.tsx
@@ -1,4 +1,4 @@
-import { type ReactNode } from "react";
+import { type ReactNode, useEffect, useRef } from "react";
 
 import { useQuery } from "@tanstack/react-query";
 
@@ -7,24 +7,42 @@ import { toast } from "react-toastify";
 import { getLeaderboardApi } from "@/api/public/get-leaderboard.api.ts";
 
 import ButtonComponent from "@/components/button/button.component";
+import ModalComponent from "@/components/modal/modal.component.tsx";
 import RibbonComponent from "@/components/ribbon/ribbon.component";
 import TableComponent from "@/components/table/table.component";
+
+import { useGameStore } from "@/stores/game.store.ts";
 
 import styles from "./game-over-modal.module.css";
 
 type Props = {
-  restartGame: () => void;
-  exitGame: () => void;
+  onRestart: () => void;
+  onExit: () => void;
 };
 
 export default function GameOverModalComponent({
-  restartGame,
-  exitGame,
+  onRestart,
+  onExit,
 }: Props): ReactNode {
   const { data, isPending, isError, error } = useQuery({
     queryKey: ["leaderboard"],
     queryFn: getLeaderboardApi,
   });
+
+  const gameState = useGameStore((state) => state.gameState);
+
+  const modalRef = useRef<HTMLDialogElement>(null);
+
+  useEffect(() => {
+    if (gameState === "over") {
+      modalRef.current?.showModal();
+    }
+  }, [gameState]);
+
+  const restartButtonClickHandler = (): void => {
+    modalRef.current?.close();
+    onRestart();
+  };
 
   if (isPending) {
     return <p className={styles.message}>Loading...</p>;
@@ -40,7 +58,11 @@ export default function GameOverModalComponent({
   }
 
   return (
-    <div className={styles["game-over-modal"]}>
+    <ModalComponent
+      ref={modalRef}
+      title="Game Over"
+      className={styles["game-over-modal"]}
+    >
       <RibbonComponent>
         <div className={styles["ribbon-content"]}>
           <div>97</div>
@@ -54,14 +76,17 @@ export default function GameOverModalComponent({
         <ButtonComponent
           color="secondary"
           className={styles.button}
-          onClick={exitGame}
+          onClick={onExit}
         >
-          EXIT
+          Exit
         </ButtonComponent>
-        <ButtonComponent className={styles.button} onClick={restartGame}>
-          RESTART
+        <ButtonComponent
+          className={styles.button}
+          onClick={restartButtonClickHandler}
+        >
+          Restart
         </ButtonComponent>
       </div>
-    </div>
+    </ModalComponent>
   );
 }
